requirements.txt
hdx-python-api